refactor: decompose polling/service.ts into 3 focused services
Extract auto-merge, CI retry, and auto-update concerns from the 1087-line
MentionPollingService into dedicated modules:

- auto-merge.ts (130 LOC) — squash-merge passing PRs authored by agent
- ci-retry.ts (212 LOC) — spawn fix sessions for CI-failed PRs with cooldown
- auto-update.ts (157 LOC) — pull origin/main and restart when idle

service.ts now focuses on its core concern (mention polling) at 701 LOC
and delegates to sub-services via start()/stop() lifecycle. Updated tests
to reference CIRetryService directly for previously-private method tests.

All 4492 tests pass. No public API changes — MentionPollingService retains
the same constructor signature and public interface.

Closes #454
