+++ server/polling/auto-merge.ts
@@ -1,0 +1,130 @@
+/**
+ * AutoMergeService — squash-merges open PRs authored by the agent that have all CI checks passing.
+ *
+ * Extracted from MentionPollingService to isolate auto-merge concerns.
+ * Runs on a 2-minute interval, scanning active polling configs for repos to check.
+ */
+
+import type { Database } from 'bun:sqlite';
+import { createLogger } from '../lib/logger';
+import { resolveFullRepo } from './github-searcher';
+
+const log = createLogger('AutoMerge');
+
+/** How often to check for mergeable PRs. */
+export const AUTO_MERGE_INTERVAL_MS = 2 * 60 * 1000; // 2 minutes
+
+export type RunGhFn = (args: string[]) => Promise<{ ok: boolean; stdout: string; stderr: string }>;
+
+export class AutoMergeService {
+    private db: Database;
+    private runGh: RunGhFn;
+    private timer: ReturnType<typeof setInterval> | null = null;
+    private running = false;
+
+    constructor(db: Database, runGh: RunGhFn) {
+        this.db = db;
+        this.runGh = runGh;
+    }
+
+    start(): void {
+        if (this.running) return;
+        this.running = true;
+        this.checkAll();
+        this.timer = setInterval(() => this.checkAll(), AUTO_MERGE_INTERVAL_MS);
+    }
+
+    stop(): void {
+        this.running = false;
+        if (this.timer) {
+            clearInterval(this.timer);
+            this.timer = null;
+        }
+    }
+
+    /**
+     * Find open PRs authored by the polling agent username that have all CI
+     * checks passing, and squash-merge them automatically.
+     */
+    async checkAll(): Promise<void> {
+        if (!this.running) return;
+
+        try {
+            // Gather unique repos from active polling configs
+            const allConfigs = this.db.query(
+                `SELECT repo, mention_username FROM mention_polling_configs WHERE status = 'active'`
+            ).all() as Array<{ repo: string; mention_username: string }>;
+
+            // Deduplicate by repo+username
+            const seen = new Set<string>();
+            const targets: Array<{ repo: string; username: string }> = [];
+            for (const c of allConfigs) {
+                const key = `${c.repo}:${c.mention_username}`;
+                if (seen.has(key)) continue;
+                seen.add(key);
+                targets.push({ repo: c.repo, username: c.mention_username });
+            }
+
+            for (const { repo, username } of targets) {
+                await this.mergeForRepo(repo, username);
+            }
+        } catch (err) {
+            log.error('Error in auto-merge loop', { error: err instanceof Error ? err.message : String(err) });
+        }
+    }
+
+    /**
+     * Auto-merge passing PRs for a specific repo authored by the given username.
+     */
+    private async mergeForRepo(repo: string, username: string): Promise<void> {
+        // List open PRs authored by the agent
+        const searchQualifier = repo.includes('/') ? `repo:${repo}` : `org:${repo}`;
+        const result = await this.runGh([
+            'api', 'search/issues',
+            '-X', 'GET',
+            '-f', `q=${searchQualifier} is:pr is:open author:${username}`,
+            '-f', 'per_page=20',
+        ]);
+
+        if (!result.ok || !result.stdout.trim()) return;
+
+        const parsed = JSON.parse(result.stdout) as { items?: Array<Record<string, unknown>> };
+        const prs = parsed.items ?? [];
+        if (prs.length === 0) return;
+
+        for (const pr of prs) {
+            const prNumber = pr.number as number;
+            const prUrl = (pr.html_url as string) ?? '';
+            const prRepo = resolveFullRepo(repo, prUrl);
+
+            // Check if all CI checks have passed
+            const statusResult = await this.runGh([
+                'pr', 'checks', String(prNumber),
+                '--repo', prRepo,
+                '--json', 'state',
+                '--jq', '[.[].state] | if length == 0 then "none" elif all(. == "SUCCESS") then "pass" else "fail" end',
+            ]);
+
+            if (!statusResult.ok || statusResult.stdout.trim() !== 'pass') {
+                continue;
+            }
+
+            // All checks pass — merge it
+            const mergeResult = await this.runGh([
+                'pr', 'merge', String(prNumber),
+                '--repo', prRepo,
+                '--squash',
+                '--delete-branch',
+            ]);
+
+            if (mergeResult.ok) {
+                log.info('Auto-merged PR', { repo: prRepo, number: prNumber });
+            } else {
+                log.debug('Failed to auto-merge PR', {
+                    repo: prRepo, number: prNumber,
+                    error: mergeResult.stderr,
+                });
+            }
+        }
+    }
+}
